move change log to separate file
A change log may contain outdated instructions that are out of sync with
other documentation in the README.  To avoid any confusion, this moves
the change log to a separate file.

The log is also converted to Markdown format so that it renders well on
Bitbucket when browsed.

Issue: CRB-3220