update build instructions for RHEL
Lifts the necessary build requirements further up in the README to
accompany the ant build instructions.  This should make it more obvious
how to get a build going quicker.

The README is expanded with detailed instructions how to configure the
default Java runtime- and development environments on RHEL and what
packages are required.  Since UiO is standardised on RHEL other systems
are not explicitly mentioned, but the same approach concerning setting
JAVA_HOME will work on all systems.

Issue: CRB-3220